[Refactor] Split stack utility across common and runtime layers
