+++ be/src/common/stack_util.h
@@ -15,9 +15,19 @@
 #pragma once
 
 #include <string>
+#include <vector>
 
 namespace starrocks {
 
 std::string get_stack_trace();
+std::vector<int> get_thread_id_list();
+std::string get_thread_name(int tid);
+bool install_stack_trace_sighandler();
+std::string get_stack_trace_for_thread(int tid, int timeout_ms);
+std::string get_stack_trace_for_threads(const std::vector<int>& tids, int timeout_ms);
+std::string get_stack_trace_for_all_threads_with_prefix(const std::string& line_prefix);
+std::string get_stack_trace_for_all_threads();
+// get all thread stack trace, and filter by function pattern
+std::string get_stack_trace_for_function(const std::string& function_pattern);
 
 } // namespace starrocks
