+++ be/src/exec/pipeline/pipeline_driver_executor.cpp
@@ -29,7 +29,6 @@
 #include "runtime/current_thread.h"
 #include "runtime/exec_env.h"
 #include "util/debug/query_trace.h"
-#include "util/stack_util.h"
 #include "util/time_guard.h"
 
 namespace starrocks::pipeline {
