+++ be/src/exec/pipeline/schedule/timeout_tasks.cpp
@@ -21,7 +21,6 @@
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/schedule/common.h"
 #include "runtime/logconfig.h"
-#include "util/stack_util.h"
 
 namespace starrocks::pipeline {
 void CheckFragmentTimeout::Run() {
