+++ be/src/runtime/CMakeLists.txt
@@ -70,6 +70,7 @@ set(RUNTIME_FILES
     sender_queue.cpp
     descriptors_ext.cpp
     exec_env.cpp
+    exception_stack.cpp
     user_function_cache.cpp
     jdbc_driver_manager.cpp
     plan_fragment_executor.cpp
