+++ be/src/runtime/diagnose_daemon.cpp
@@ -17,7 +17,7 @@
 #include <fmt/format.h>
 
 #include "common/config_exec_flow_fwd.h"
-#include "util/stack_util.h"
+#include "common/stack_util.h"
 
 namespace starrocks {
