+++ be/src/storage/convert_helper.cpp
@@ -36,7 +36,6 @@
 #include "types/percentile_value.h"
 #include "types/storage_type_traits.h"
 #include "types/timestamp_value.h"
-#include "util/stack_util.h"
 
 namespace starrocks {
