+++ be/src/util/CMakeLists.txt
@@ -28,7 +28,6 @@ set(UTIL_FILES
 # TODO: not supported on RHEL 5
 # perf-counters.cpp
   static_asserts.cpp
-  stack_util.cpp
   time_guard.cpp
   system_metrics.cpp
 # coding_util.cpp
